refactor : 오타 수정, 엔티티명 직관적이게 변경 및 패키지 이동
